Extracted report.
